src/lrg/requirements.txt
asgiref==3.8.1
boto3==1.36.21
botocore==1.36.21
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-livereload-server==0.5.1
Faker==36.1.1
geosky==1.0.9
geonamescache==2.0.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==24.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
tornado==6.4.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.9.0
